requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
langchain==1.2.12
langchain-core==1.2.18
langchain-ollama==1.0.1
langchain-openai==1.1.11
langgraph==1.1.1
langgraph-prebuilt==1.0.8
python-dotenv==1.2.2
pydantic==2.12.5
